[MOV] website_twitter: split the module files
